internal/grpc: separate grpcServer and xdsHandler
This PR separates the grpcServer, which implements the xDS grpc methods,
from a new type, the xdsHandler, which is concerned with serving xds
traffic generically.
